refactor(openomni,server): move tool system to execution-runtime module
Move NativeTool, ToolProvider, executor, builtins, system/agent providers, and
workspace-path utils from apps/server/src/tool/ to
packages/openomni/src/execution-runtime/tool/ so workers can consume them
without depending on server. MCP provider stays in server.
Update all import sites to use @openomni/openomni.
