+++ apps/server/src/bootstrap/index.ts
@@ -2,14 +2,12 @@ import { mkdirSync } from "node:fs";
 import { dirname } from "node:path";
 import type { Adapter } from "@openomni/protocol";
 import { initialize } from "@openomni/session";
-import { IngressEngine } from "@openomni/openomni";
+import { AgentToolProvider, IngressEngine, SystemToolProvider } from "@openomni/openomni";
 import { loadConfig } from "../config";
 import { createMessageHandler } from "../handler/conversation";
 import { createExecutionCoordinator } from "../execution/coordinator";
 import { createRouter } from "../server/routes";
-import { AgentToolProvider } from "../tool/agent";
 import { McpToolProvider } from "../tool/mcp";
-import { SystemToolProvider } from "../tool/system";
 import { createChannelAdapters } from "./channels";
 import { connectMcpServers } from "./mcp";
 import { resolveModel } from "./providers";
