--- apps/server/src/tool/agent/index.ts
@@ -1,2 +1,0 @@
-export { AgentToolProvider } from "./provider";
-export { createSubagentTool } from "./tools/subagent";
