--- apps/server/src/tool/index.ts
@@ -1,15 +1,0 @@
-export { AgentToolProvider } from "./agent";
-export { Tool, defineTool, resolveMeta } from "./define";
-export { McpToolProvider } from "./mcp";
-export { createToolExecutor } from "./executor";
-export { SystemToolProvider } from "./system";
-export type { ToolExecutorContext } from "./executor";
-export type {
-  NativeTool,
-  ToolCategory,
-  ToolExecutorConfig,
-  ToolMetaValue,
-  ToolProvider,
-  ToolRiskTier,
-  ToolSource,
-} from "./types";
