+++ apps/server/src/tool/mcp/provider.ts
@@ -1,6 +1,6 @@
 import { McpClient, type McpServerConfig } from "@openomni/agent";
 import type { Tool } from "@openomni/protocol";
-import type { NativeTool, ToolCategory, ToolProvider } from "../types";
+import type { NativeTool, ToolCategory, ToolProvider } from "@openomni/openomni";
 
 export class McpToolProvider implements ToolProvider {
   readonly name = "mcp";
