--- apps/server/src/tool/system/index.ts
@@ -1,7 +1,0 @@
-export { SystemToolProvider } from "./provider";
-export { bashTool } from "../builtins/bash";
-export { createEditTool } from "../builtins/edit";
-export { createGlobTool } from "../builtins/glob";
-export { createGrepTool } from "../builtins/grep";
-export { createReadTool } from "../builtins/read";
-export { createWriteTool } from "../builtins/write";
