+++ packages/openomni/src/execution-runtime/filesystem/workspace-path.ts
