+++ packages/openomni/src/execution-runtime/index.ts
@@ -1,2 +1,20 @@
 export { buildWorkerMiddleware } from "./middleware.js";
 export type { WorkerMiddlewareConfig } from "./middleware.js";
+export {
+  AgentToolProvider,
+  SystemToolProvider,
+  Tool,
+  createToolExecutor,
+  defineTool,
+  resolveMeta,
+} from "./tool/index.js";
+export type {
+  NativeTool,
+  ToolCategory,
+  ToolExecutorConfig,
+  ToolExecutorContext,
+  ToolMetaValue,
+  ToolProvider,
+  ToolRiskTier,
+  ToolSource,
+} from "./tool/index.js";
