+++ packages/openomni/src/execution-runtime/tool/builtins/bash-classify.ts
