+++ packages/openomni/src/execution-runtime/tool/builtins/bash-prompt.ts
