+++ packages/openomni/src/execution-runtime/tool/builtins/edit-prompt.ts
