+++ packages/openomni/src/execution-runtime/tool/builtins/grep-prompt.ts
