+++ packages/openomni/src/execution-runtime/tool/builtins/read-prompt.ts
