+++ packages/openomni/src/execution-runtime/tool/builtins/write-prompt.ts
