+++ packages/openomni/src/execution-runtime/tool/shared/input.ts
