+++ packages/openomni/src/execution-runtime/tool/shared/result.ts
